fix GD::Simple->fontMetrics (GH #37)
The GD::Simple->fontMetrics() function doesn't seem to be coded in such a
way that it does what the documentation says. The documentation at
https://metacpan.org/pod/GD::Simple says that the metrics are defined in
terms of the rendering of the letters m, j, and d, but the code at
https://metacpan.org/release/GD/source/lib/GD/Simple.pm actually only
does test renderings of m and j, never d.

The documentation says that the lineheight is defined as
"the distance from the bottom of the 'j' to the top of the 'd'".
This sounds like the intention is to return the height of the string
'jd'. This would automatically be greater than the xheight. Actually
what it seems to calculate is the whitespace in the middle of "m\nm",
which is much smaller.

Signed-off-by: Reini Urban <rurban@cpan.org>
This could break old code, but the code is now correct
and matches the docs.